Added Ghost subscriber capability
Added the new Ghost subscriber option to the bottom of each post page
and also to the subscribe button at the bottom of the navigation
element.  Both will only show if you’ve enabled the beta feature.  More
on subscribers here:
https://help.ghost.org/hc/en-us/articles/224089787-Subscribers-Beta

The button at the bottom of the posts page could use a little CSS love.
 I’m no expert, maybe someone else could have a look?