Improve navigation landmark announcements for screen readers.
Navigations were spoken multiple times because a nav implies a navigation landmark, and a role "navigation" causes another navigation landmark to be created inside of it.

Also added labels to the navigations so they are distinguishable from one another. Note that because of the navigation role, screen readers already speak the word "navigation" in a localized form, the labels only contain the rest, without the word "navigation".